+++ Converter/BoolToStringConverter.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Data;
 
-namespace WpfApp.Converter
+namespace WpfLib.Converter
 {
     /// <summary>
     /// Takes a bool and outputs a string according to the string value of ConverterParameter. The ConverterParameter should be used as "true|false" so if you want to convert 'true' to 'yes' and 'false' to 'no' set ConverterParameter to "yes|no".